requirements.txt
google-generativeai
flask
python-dotenv
Pillow